test_logger.py 파일도 변경된 파일 구조(core/loggers/ 하위 모듈)에 맞춰 독립적인 파일들로 분리했습니다.
새롭게 분리된 단위 테스트 파일들은 다음과 같습니다 (tests/unit_test/core/loggers/ 경로 생성됨):

test_app_logger.py: Logger 메인 클래스의 동작, 파일 회전(rotation), 예외 처리, 정리 로직 등을 테스트.
test_size_time_rotating_file_handler.py: SizeTimeRotatingFileHandler의 파일 개수 관리 및 인덱스 동작 테스트.
test_cache_event_logger.py: CacheEventLogger 및 관련된 로깅 동작들(현재가/OHLCV히트 및 미스 로그, 메모리LRU/LFU축출 동작 검증 등) 테스트.
test_streaming_event_logger.py: StreamingEventLogger 및 웹소켓 워치독(task_count,subscribe/unsubscribe, reconnect 등) 테스트.
또한 기존의 tests/unit_test/core/test_logger.py 파일은 core/logger.py에 여전히 남아있는 팩토리 함수들(get_strategy_logger, get_performance_logger 등)에 대한 테스트만 남기도록 슬림하게 변경했습니다.

pytest를 사용해(파이썬 3.10 환경) 검증까지 완료하여 모든 71개 테스트 항목들이 정상적으로 통과됨을 확인했습니다!
